chore: Remove unused Home screen components
