+++ src/app/Components/Home/ArtistRails/RecommendedArtistsRail.tsx
@@ -1,5 +1,5 @@
 import { ActionType, OwnerType } from "@artsy/cohesion"
-import { Spacer, Flex, Spinner } from "@artsy/palette-mobile"
+import { Flex, Spacer, Spinner } from "@artsy/palette-mobile"
 import { ArtistCard_artist$data } from "__generated__/ArtistCard_artist.graphql"
 import { RecommendedArtistsRailFollowMutation } from "__generated__/RecommendedArtistsRailFollowMutation.graphql"
 import { RecommendedArtistsRail_me$data } from "__generated__/RecommendedArtistsRail_me.graphql"
@@ -7,7 +7,7 @@ import { CardRailFlatList } from "app/Components/Home/CardRailFlatList"
 import { SectionTitle } from "app/Components/SectionTitle"
 import { defaultArtistVariables } from "app/Scenes/Artist/Artist"
 import { RailScrollProps } from "app/Scenes/Home/Components/types"
-import HomeAnalytics from "app/Scenes/Home/homeAnalytics"
+import HomeAnalytics from "app/Scenes/HomeView/helpers/homeAnalytics"
 import { getRelayEnvironment } from "app/system/relay/defaultEnvironment"
 import { extractNodes } from "app/utils/extractNodes"
 import React, { memo, useImperativeHandle, useRef, useState } from "react"
