+++ src/app/Scenes/HomeView/helpers/homeAnalytics.tests.ts
