+++ src/app/Scenes/HomeView/helpers/homeAnalytics.ts
